Fix minor review issues
